backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
openai
pypdf
sqlalchemy
psycopg2-binary
alembic
PyJWT
